update ch7 pp6 solution

+++ Ch07_Basic_Types/ch7_prog_proj_06.c
@@ -11,12 +11,12 @@
 
 int main(void)
 {
-	printf("Size of int is %u\n", sizeof(int));
-	printf("Size of short is %u\n", sizeof(short));
-	printf("Size of long is %u\n", sizeof(long));
-	printf("Size of float is %u\n", sizeof(float));
-	printf("Size of double is %u\n", sizeof(double));
-	printf("Size of long double is %u\n", sizeof(long double));
+	printf("Size of int is %lu\n", (unsigned long) sizeof(int));
+	printf("Size of short is %lu\n", (unsigned long) sizeof(short));
+	printf("Size of long is %lu\n", (unsigned long) sizeof(long));
+	printf("Size of float is %lu\n", (unsigned long) sizeof(float));
+	printf("Size of double is %lu\n", (unsigned long) sizeof(double));
+	printf("Size of long double is %lu\n", (unsigned long) sizeof(long double));
 
 	return 0;
 }
